pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "naics-mcp-server"
version = "0.1.0"
description = "NAICS MCP Server - Intelligent industry classification service for NAICS 2022"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
keywords = ["naics", "industry", "classification", "mcp", "llm"]
authors = [
    { name = "NAICS MCP Team" }
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "mcp>=1.0.0",
    "duckdb>=0.9.0",
    "sentence-transformers>=2.2.0",
    "numpy>=1.24.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.0.280",
    "mypy>=1.0.0",
]

[project.scripts]
naics-mcp = "naics_mcp_server.cli:main"
naics-mcp-server = "naics_mcp_server.server:main"

[project.urls]
Documentation = "https://github.com/your-org/naics-mcp-server#readme"
Issues = "https://github.com/your-org/naics-mcp-server/issues"
Source = "https://github.com/your-org/naics-mcp-server"

[tool.hatch.build.targets.wheel]
packages = ["naics_mcp_server"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I", "N", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
